+++ raycaster/Vector2.h
@@ -4,10 +4,10 @@
 class Vector2 {
 private:
 public:
-	float x, y;
+	double x, y;
 	Vector2()
 		: x(0), y(0) {}
-	Vector2(float x, float y)
+	Vector2(double x, double y)
 		: x(x), y(y) {}
 	Vector2(const Vector2& v)
 		: x(v.x), y(v.y) {}
@@ -17,27 +17,27 @@ public:
 	{ return Vector2(x + that.x, y + that.y); }
 	Vector2 mul(const Vector2 &that)
 	{ return Vector2(x*that.x, y*that.y); }
-	Vector2 mul(float f)
+	Vector2 mul(double f)
 	{ return Vector2(x*f, y*f); }
-	float length()
+	double length()
 	{ return sqrt(x*x + y*y); }
 	Vector2 norm()
-	{ float l = length();
+	{ double l = length();
 	if (l == 0) return Vector2(0, 0);
 	return Vector2(x/l, y/l); }
-	Vector2 scale(float f)
+	Vector2 scale(double f)
 	{ return Vector2(x*f, y*f); }
-	float dot(const Vector2 &that)
+	double dot(const Vector2 &that)
 	{ return x*that.x + y*that.y; }
 	Vector2 rot90()
 	{ return Vector2(-y, x); }
 	Vector2 rot_minus90()
 	{ return Vector2(y, -x); }
-	Vector2 lerp(const Vector2 &that, float t)
+	Vector2 lerp(const Vector2 &that, double t)
 	{ return Vector2(x + t * (that.x - x), y + t * (that.y - y)); }
 };
 
-inline Vector2 from_angle(float angle)
+inline Vector2 from_angle(double angle)
 { return Vector2(cos(angle), sin(angle)); }
 
 inline Vector2 zero()
